reworked project structure to follow barrel pattern
