+++ src/models/index.js
@@ -1,3 +1,5 @@
-import { userProfiles, documents, userLogins } from './dms';
+import { documents } from './documents';
+import { userLogins } from './userLogins';
+import { userProfiles } from './userProfiles';
 
 export { userProfiles, documents, userLogins };
