+++ src/routes/usersRoute.js
@@ -1,16 +1,16 @@
 import express from 'express';
 import { errorHandlers } from '../middleware';
-import { userProfilesService, userLoginsService } from '../services';
+import { usersService } from '../services';
 
 const usersRouter = express.Router();
 const { generic } = errorHandlers;
 
-usersRouter.get('/users', userProfilesService.fetchActiveProfiles);
-usersRouter.get('/users/:id', userProfilesService.fetchProfile);
-usersRouter.post('/users', userProfilesService.createProfile);
-usersRouter.put('/users/:id', userProfilesService.updateProfile);
-usersRouter.delete('/users/:id', userProfilesService.deleteProfile);
-usersRouter.post('/users/login', userLoginsService.logIn);
+usersRouter.get('/users', usersService.fetchActiveProfiles);
+usersRouter.get('/users/:id', usersService.fetchProfile);
+usersRouter.post('/users', usersService.createProfile);
+usersRouter.put('/users/:id', usersService.updateProfile);
+usersRouter.delete('/users/:id', usersService.deleteProfile);
+usersRouter.post('/users/login', usersService.logIn);
 usersRouter.use('/users', (err, req, res, next) => {
   // if response has already started streaming and error occurs, pass it to Express
   // default error handler - it will close the connection and fail the request.
